merge 20160620

+++ general.sh
@@ -132,19 +132,21 @@ if [ -d "$SOURCES/$2/$GITHUBSUBDIR" ]; then
 	if [[ "$3" != "" ]] && [[ "$bar_1" == "$localbar" || "$bar_2" == "$localbar" ]] || [[ "$3" == "" && "$bar_3" == "$localbar" ]] || [[ $bar_1 == "" && $bar_2 == "" ]]; then
 		display_alert "... you have latest sources" "$2 $3" "info"
 	else		
-		display_alert "... your sources are outdated - creating new shallow clone" "$2 $3" "info"
-		if [[ -z "$GITHUBSUBDIR" ]]; then 
-			rm -rf $SOURCES/$2".old"
-			mv $SOURCES/$2 $SOURCES/$2".old" 
-		else
-			rm -rf $SOURCES/$2/$GITHUBSUBDIR".old"
-			mv $SOURCES/$2/$GITHUBSUBDIR $SOURCES/$2/$GITHUBSUBDIR".old" 
-		fi
-		
-		if [[ -n $3 && -n "$(git ls-remote $1 | grep "$tag")" ]]; then
-			git clone -n $1 $SOURCES/$2/$GITHUBSUBDIR -b $3 --depth 1 || git clone -n $1 $SOURCES/$2/$GITHUBSUBDIR -b $3
-		else
-			git clone -n $1 $SOURCES/$2/$GITHUBSUBDIR --depth 1
+		if [ $DEBUG_MODE != yes ]; then
+			display_alert "... your sources are outdated - creating new shallow clone" "$2 $3" "info"
+			if [[ -z "$GITHUBSUBDIR" ]]; then 
+				rm -rf $SOURCES/$2".old"
+				mv $SOURCES/$2 $SOURCES/$2".old" 
+			else
+				rm -rf $SOURCES/$2/$GITHUBSUBDIR".old"
+				mv $SOURCES/$2/$GITHUBSUBDIR $SOURCES/$2/$GITHUBSUBDIR".old" 
+			fi
+			
+			if [[ -n $3 && -n "$(git ls-remote $1 | grep "$tag")" ]]; then
+				git clone -n $1 $SOURCES/$2/$GITHUBSUBDIR -b $3 --depth 1 || git clone -n $1 $SOURCES/$2/$GITHUBSUBDIR -b $3
+			else
+				git clone -n $1 $SOURCES/$2/$GITHUBSUBDIR --depth 1
+			fi
 		fi
 		cd $SOURCES/$2/$GITHUBSUBDIR
 		git checkout -q
@@ -179,7 +181,8 @@ display_alert()
 # log function parameters to install.log
 echo "Displaying message: $@" >> $DEST/debug/install.log
 
-[[ -n $2 ]] && local tmp="[\e[0;33m $2 \x1B[0m]"
+local tmp=""
+[[ -n $2 ]] && tmp="[\e[0;33m $2 \x1B[0m]"
 
 case $3 in
 	err)
@@ -345,6 +348,8 @@ prepare_host() {
 		apt-key adv --keyserver keys.gnupg.net --recv-keys 9E3E53F19C7DE460
 	fi
 
+	if [[ $codename == xenial ]]; then hostdeps="$hostdeps systemd-container"; fi
+
 	# Deboostrap in trusty breaks due too old debootstrap. We are installing Xenial package
 	local debootstrap_version=$(dpkg-query -W -f='${Version}\n' debootstrap | cut -f1 -d'+')
 	local debootstrap_minimal="1.0.78"
@@ -381,7 +386,7 @@ prepare_host() {
 	test -e /proc/sys/fs/binfmt_misc/qemu-aarch64 || update-binfmts --enable qemu-aarch64
 
 	# create directory structure
-	mkdir -p $SOURCES $DEST/debug $CACHEDIR/rootfs $SRC/userpatches/overlay $SRC/toolchains
+	mkdir -p $SOURCES $DEST/debug $CACHEDIR/rootfs $SRC/userpatches/overlay $SRC/toolchains $SRC/userpatches/patch
 	find $SRC/lib/patch -type d ! -name . | sed "s%lib/patch%userpatches%" | xargs mkdir -p
 
 	# download external Linaro compiler and missing special dependencies since they are needed for certain sources
